requirements.txt
asgiref==3.5.2
cachetools==5.2.0
Django==3.2.16
django-allauth==0.41.0
django-crispy-forms==1.14.0
google-api-core==2.11.0
google-auth==2.15.0
google-cloud-vision==3.2.0
googleapis-common-protos==1.57.0
grpcio==1.51.1
grpcio-status==1.51.1
oauthlib==3.2.2
Pillow==9.3.0
proto-plus==1.22.1
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
python3-openid==3.2.0
pytz==2022.6
requests-oauthlib==1.3.1
rsa==4.9
sqlparse==0.4.3
